requirements.txt
# IMPORTANT: This is only used for the Streamlit app deployment
# to Streamlit Community Cloud, since it doesn't have support for uv.
# USE PYPROJECT.TOML AND UV.LOCK FOR YOUR OWN INSTALL!

httpx~=0.27.2
pydantic ~=2.10.1
streamlit~=1.40.1
numexpr>=2.8.8
cohere>=5.0.0  # Required for v2 API support
pinecone-client>=3.0.0
